Decouple middleware and mixins
