+++ src/core/interfaces.d.ts
@@ -1,3 +1,5 @@
+/* tslint:disable:interface-name */
+import { Bundle, Messages } from '../i18n/i18n';
 import { Destroyable } from '../core/Destroyable';
 import { Evented, EventType, EventObject } from '../core/Evented';
 import Map from '../shim/Map';
@@ -113,6 +115,20 @@ export interface Projection {
 }
 
 export type SupportedClassName = string | null | undefined | boolean;
+
+export type ClassNames = {
+	[key: string]: string;
+};
+
+export interface Theme {
+	[key: string]: object;
+}
+
+export interface Classes {
+	[widgetKey: string]: {
+		[classKey: string]: SupportedClassName[];
+	};
+}
 
 export type DeferredVirtualProperties = (inserted: boolean) => VNodeProperties;
 
@@ -601,3 +617,55 @@ export interface AfterRender {
 export interface BeforeProperties<P = any> {
 	(properties: P): P;
 }
+
+export interface LocaleData {
+	/**
+	 * The locale for the widget. If not specified, then the root locale (as determined by `@dojo/i18n`) is assumed.
+	 * If specified, the widget's node will have a `lang` property set to the locale.
+	 */
+	locale?: string;
+
+	/**
+	 * An optional flag indicating the widget's text direction. If `true`, then the underlying node's `dir`
+	 * property is set to "rtl". If it is `false`, then the `dir` property is set to "ltr". Otherwise, the property
+	 * is not set.
+	 */
+	rtl?: boolean;
+}
+
+export interface I18nProperties extends LocaleData {
+	/**
+	 * An optional override for the bundle passed to the `localizeBundle`. If the override contains a `messages` object,
+	 * then it will completely replace the underlying bundle. Otherwise, a new bundle will be created with the additional
+	 * locale loaders.
+	 */
+	i18nBundle?: Bundle<Messages> | Map<Bundle<Messages>, Bundle<Messages>>;
+}
+
+export type LocalizedMessages<T extends Messages> = {
+	/**
+	 * Indicates whether the messages are placeholders while waiting for the actual localized messages to load.
+	 * This is always `false` if the associated bundle does not list any supported locales.
+	 */
+	readonly isPlaceholder: boolean;
+
+	/**
+	 * Formats an ICU-formatted message template for the represented bundle.
+	 *
+	 * @param key
+	 * The message key.
+	 *
+	 * @param options
+	 * The values to pass to the formatter.
+	 *
+	 * @return
+	 * The formatted string.
+	 */
+	format(key: string, options?: any): string;
+
+	/**
+	 * The localized messages if available, or either the default messages or a blank bundle depending on the
+	 * call signature for `localizeBundle`.
+	 */
+	readonly messages: T;
+};
